# lib/api-zod/src/generated/types/generalAssignRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneralAssignRequestSlot } from "./generalAssignRequestSlot";

export interface GeneralAssignRequest {
  campaignId: number;
  slot: GeneralAssignRequestSlot;
}