# lib/api-zod/src/generated/types/generalCosts.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeneralCosts {
  /** 抽取成本：5% 國庫 */
  drawMoney: number;
  /** 抽取成本：5% 可用生產力 */
  drawProduction: number;
}