# lib/api-zod/src/generated/types/generalDrawResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { General } from "./general";
import type { GeneralDrawResponseSpent } from "./generalDrawResponseSpent";

export interface GeneralDrawResponse {
  general: General;
  /** 本次抽取實扣 */
  spent: GeneralDrawResponseSpent;
}