# lib/api-zod/src/generated/types/generalSkill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneralSkillEffect } from "./generalSkillEffect";

export interface GeneralSkill {
  /** 技能名稱（AI 生成） */
  name: string;
  /** 技能描述（敘事用） */
  description: string;
  /** 效果類型 */
  effect: GeneralSkillEffect;
  /** 效果幅度 %（0–10） */
  bonusPct: number;
  /** 解鎖品級（1/2/4） */
  unlockGrade: number;
}