# lib/api-zod/src/generated/types/generalUpgradeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { General } from "./general";
import type { GeneralUpgradeResponseSpent } from "./generalUpgradeResponseSpent";

export interface GeneralUpgradeResponse {
  general: General;
  /** 本次擲骰結果（失敗照扣成本） */
  success: boolean;
  /** 本次成功率 % */
  successPct: number;
  /** 本次升階實扣 */
  spent: GeneralUpgradeResponseSpent;
}