# lib/api-zod/src/generated/types/generalsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { General } from "./general";
import type { GeneralAssignmentOption } from "./generalAssignmentOption";
import type { GeneralCosts } from "./generalCosts";
import type { GeneralsQuota } from "./generalsQuota";

export interface GeneralsListResponse {
  generals: General[];
  quota: GeneralsQuota;
  costs: GeneralCosts;
  /** 該國進行中戰役的所有軍團（可指派目標） */
  assignmentOptions: GeneralAssignmentOption[];
}