# lib/api-zod/src/generated/types/generalsQuota.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeneralsQuota {
  /** 在營上限（8） */
  cap: number;
  /** 在營數 */
  recruited: number;
  /** 候選數 */
  candidates: number;
  /** 本回合已抽取 */
  drawnThisTurn: boolean;
}